chore(app): setup postgresql database
add database to persist data
add intervention API endpoints
change line break from CRLF to LF
change the folder structure
add function to create and drop tables

[Delivers #16253287]
