+++ package.json
@@ -12,7 +12,9 @@
     "coverage": "nyc npm test && nyc report --reporter=lcov --reporter=text --reporter=lcovonly || coveralls",
     "local": "nyc --reporter=text --reporter=html mocha ./app/tests/test.js --require babel-polyfill --require babel-register --exit",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
-    "lint": "eslint ./app"
+    "lint": "eslint ./app",
+    "create-table": "node build/models/createTable",
+    "drop-table": "node build/models/dropTable"
   },
   "repository": {
     "type": "git",
